Add protected config attribute and support for redirecting via session variable
Redirecting on the config attribute 'authenticated' value (routes in laravel)
proved to be limiting since there are needs to use shib on more than one
route. Added the protected array to be used with custom middleware in
laravel to detect if a given route is within the protected array.

Default idpAuthenticate method would only redirect to authenticated
config but updated to also listen on a session variable called shibboleth_redirect_to
that can be set from custom middleware.